Warn users if VS2017 environment variable not set
There is an issue with the VS2017 RC installer whereby it doesn't set
the VS150COMNTOOLS environment variable. This environment variable has
been a long standing convention for Visual Studio and hopefully will be
addressed in the official release of VS2017.

Also tidied up some of the help discussing valid input for the
$VisualStudioVersion parameter.