Add support for -EnableDocumentFunction to Convert-Xml
The Convert-Xml cmdlet performs XML transforms using the XslCompiledTransform
class.  That class defaults to not allowing use of the 'document()' function,
which allows a transform to reference data in a local or remote file.

This commit adds a parameter, -EnableDocumentFunction, to the Convert-Xml
cmdlet, making it possible to perform transforms that require document().